+++ resources/compiledJS/game/start_game.js
@@ -390,16 +390,14 @@ function initGame() {
     state.menuOpen = true;
     state.titleScreen = true;
     gotoMainMenu(true);
-    if (DEVMODE) {
-        renderMinimap(maps[currentMap]);
-        renderAreaMap(maps[currentMap]);
-        createStaticMap();
-        modifyCanvas(true);
-    }
+    renderMinimap(maps[currentMap]);
+    renderAreaMap(maps[currentMap]);
+    createStaticMap();
+    modifyCanvas(true);
     tooltip(document.querySelector(".invScrb"), `${lang["setting_hotkey_inv"]} [${settings["hotkey_inv"]}]`);
     tooltip(document.querySelector(".chaScrb"), `${lang["setting_hotkey_char"]} [${settings["hotkey_char"]}]`);
     tooltip(document.querySelector(".perScrb"), `${lang["setting_hotkey_perk"]} [${settings["hotkey_perk"]}]`);
     tooltip(document.querySelector(".escScrb"), `${lang["open_menu"]} [ESCAPE]`);
 }
-initGame();
+document.addEventListener("DOMContentLoaded", initGame);
 //# sourceMappingURL=start_game.js.map