+++ resources/compiledJS/ui/hotbar.js
@@ -226,11 +226,17 @@ function generateEffects() {
     const effects = document.querySelector(".playerEffects");
     effects.textContent = "";
     player.statusEffects.forEach((effect) => {
+        const wrapper = document.createElement("div");
         const img = document.createElement("img");
+        const bg = document.createElement("div");
+        wrapper.classList.add("status");
         img.classList.add("status");
+        bg.classList.add("status-bg");
+        bg.style.height = `${Math.min((effect.last.current / effect.last.total) * 100, 100)}%`;
         img.src = effect.icon;
-        tooltip(img, statTT(effect));
-        effects.append(img);
+        wrapper.append(img, bg);
+        tooltip(wrapper, statTT(effect));
+        effects.append(wrapper);
     });
 }
 function keyIncludesAbility(key) {